app/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
pymupdf
Pillow
imagehash
python-multipart
